mcp-examples/11-tech-debt/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tech-debt-mcp"
version = "0.1.0"
description = "Technical Debt Analyzer MCP Server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "sonarqube-api>=1.3.0",
    "lizard>=1.17.10",
    "radon>=6.0.1",
    "git-python>=1.0.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.2", "black>=23.9.1", "ruff>=0.0.292"]
